requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
attrs==24.2.0
audio-recorder-streamlit==0.0.10
av==14.0.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
entrypoints==0.4
exceptiongroup==1.2.2
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
Jinja2==3.1.4
jiter==0.8.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
livekit==0.18.1
livekit-agents==0.12.1
livekit-api==0.8.0
livekit-plugins-openai==0.10.9
livekit-protocol==0.7.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
narwhals==1.15.2
numpy==2.0.2
openai==1.57.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.1
protobuf==5.29.1
psutil==5.9.8
pyarrow==18.1.0
PyAudio==0.2.14
pycares==4.5.0
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
SpeechRecognition==3.11.0
streamlit==1.40.2
streamlit-float==0.3.5
tenacity==9.0.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
types-protobuf==4.25.0.20240417
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchfiles==0.24.0
yarl==1.18.3
